fix plugin node warning
we are providing the node name parameter via launch file.
However, we only want that parameter to affect the remap manager
node, not the node created by each plugin. The issue is that,
once the node name is provided via launch file (that is,
embedded into the command line parameters), this is inevitably
propagated downstream to all the node sharing the same context
as the main one (that is, all the plugin nodes). Here, we are
proposing a solution to fix this aspect, taking inspiration
from the yolo_ros_onnx package.